Introduce Hax_ExprLongLong, Hax_GetLong and Hax_GetLongLong
In the 64-bit world, int is no longer sufficient integer value.

Switch what applicable to wider integer functions.

+++ haxInt.h
@@ -640,13 +640,13 @@ extern void		HaxExpandParseValue (ParseValue *pvPtr,
 			    int needed);
 extern int		HaxFindElement (Hax_Interp *interp,
 			    char *list, char **elementPtr, char **nextPtr,
-			    int *sizePtr, int *bracePtr);
+			    long int *sizePtr, int *bracePtr);
 extern Proc *		HaxFindProc (Interp *iPtr,
 			    char *procName);
 extern int		HaxGetFrame (Hax_Interp *interp,
 			    char *string, CallFrame **framePtrPtr);
 extern int		HaxGetListIndex (Hax_Interp *interp,
-			    char *string, int *indexPtr);
+			    char *string, long int *indexPtr);
 extern int		HaxGetOpenFile (Hax_Interp *interp,
 			    char *string, OpenFile **filePtrPtr);
 extern Proc *		HaxIsProc (Command *cmdPtr);
